[refactor] 어드민 페이지 컴포넌트 분리 (CSR)
