+++ components/admin/admin-header.tsx
@@ -1,0 +1,7 @@
+export default function AdminHeader() {
+  return (
+    <div className="flex justify-between items-center mb-6">
+      <h1 className="text-3xl font-bold">관리자 페이지</h1>
+    </div>
+  );
+}
